Update index.js files
